Numeric Key mode for quick key annotation
toggling the 'numeric key' button auto-populates the label with a number, incrementing each time a label is placed. Hand-edit the number if you don't like the current value.

Also, set the initial label size to a relative percentage of the image width, for a better ballpark guess at a legible key label.